query modifier improvements
